Flyttet tabell-logikk til elementUtils, og flyttet visningsvariant-logikk i egen fil
